Fixed a few syntax bugs with QuadraticOrder, added stubs for Quadratic tests, and began writing fpReps.hpp
Most of the changes with QuadraticOrder were simple syntactical fixes, however
there are a couple of issues that should be addressed.  First of all, the
QuadraticOrder.hpp has a member variable of type  qo_reduce_plain<T>, however
there is no such type.  There is only qo_reduce_plain_real<T>  and
qo_reduce_plain_imag<T>.  I recommend building an abstract superclass
qo_reduce_plain<T> so that we don't need to know if the order is real or
imaginary when allocating space for the reduction algorithm.  For the moment
I have temporarily commented out the problem member variable.

Secondly, there seems to be a problem with a circular dependancy in qo_cube_mulsqr
I separated the declaration from the implementation, thus creating a new, very short
file, in hopes to clarify this bug, but I am still unable to determine the exact
source of the problem, or any potential fixes.  It seems that using the friend operations
on the ideals themselves, as opposed to their member functions is why this implementation
is failing, while the others are fine.

Thirdly, I had wanted to write a test class for QuadraticOrder, however, as it
still does not compile, this class remains a stub.

Finally, I began writing an implementation for (f,p) representations.  The code
in fpReps.hpp is probably still full of problems, but until the base classes are
working, there is little I can do for it.

Note that I have not completed any debugging for QuadraticNumber, or QuadraticIdealBase.

-Reginald Lybbert